Using float angles instead of int

+++ Servo_HighLevelApp/Servo_HighLevelApp/servo.c
@@ -56,7 +56,7 @@ int SERVO_Init(struct SERVO_Config* config, struct _SERVO_State** state)
 	return 0;
 }
 
-int SERVO_SetAngle(struct _SERVO_State* servo, int angle)
+int SERVO_SetAngle(struct _SERVO_State* servo, float angle)
 {
 	if (NULL == servo)
 	{
